DMT-18 Introduce donutState.js
Decouple logic for initializing the state of the donut-chart into a separate class and expose an interface in order to reduce coupling.
