+++ src/donutState.js
@@ -1,79 +1,9 @@
 /**
- *
- * @param {Spotfire.Size} size
- * @param {any[]} data
- * @param {Spotfire.Mod} mod
- */
-import * as d3 from "d3";
-
-function donutChart(size, data, mod) {
-    // Added a constant to remove the magic numbers within the width, height and radius calculations.
-    const sizeModifier = 40;
-    // D3 animation duration used for svg shapes
-    const animationDuration = 300;
-
-    const width = size.width - sizeModifier;
-    const height = size.height - sizeModifier;
-    const radius = Math.min(width, height) / 2 - sizeModifier;
-
-    d3.select("#mod-container svg").attr("width", width).attr("height", height);
-
-    const svg = d3.select("#mod-container svg g").attr("transform", `translate(${width / 2}, ${height / 2})`);
-
-    const pie = d3.pie().value((d) => d.value);
-
-    const arc = d3
-        .arc()
-        .padAngle(0.1 / data.length)
-        .innerRadius(radius * 0.5)
-        .outerRadius(radius);
-
-    // Join new data
-    const sectors = svg.selectAll("path").data(pie(data), (d) => {
-        return d.data.id;
-    });
-
-    let newSectors = sectors
-        .enter()
-        .append("path")
-        .on("click", function (d) {
-            d.data.mark();
-        })
-        .on("mouseenter", function (d) {
-            mod.controls.tooltip.show(d.data.tooltip());
-        })
-        .on("mouseleave", function (d) {
-            mod.controls.tooltip.hide();
-        })
-        .attr("fill", (d) => "transparent");
-
-    sectors.merge(newSectors)
-        .transition()
-        .duration(animationDuration)
-        .attr("fill", (d) => d.data.color)
-        .attrTween("d", tweenArc)
-        .attr("stroke", "none");
-
-    function tweenArc(elem) {
-        let prevValue = this.__prev || {};
-        let newValue = elem;
-        this.__prev = elem;
-
-        var i = d3.interpolate(prevValue, newValue);
-
-        return function (value) {
-            return arc(i(value));
-        };
-    }
-
-    sectors.exit().transition().duration(animationDuration).attr("fill", "transparent").remove();
-}
-
-/**
  * Render the visualization
  * @param {Spotfire.Mod} mod API
+ * @return donutState
  */
-export async function render(mod) {
+export async function createDonutState(mod) {
     /**
      * Initialize dataView, size, and context based on the mod API
      * @param {Spotfire.DataView} dataView
@@ -132,10 +62,19 @@ export async function render(mod) {
             }
         };
     });
+    /**
+     * @typedef {donutState} donutState containing dataView, size, data[], modControls, context
+     */
+    let donutState = {
+        data: data,
+        size: size,
+        dataView: dataView,
+        modControls: mod.controls,
+        context: context,
+        clearMarking: () => dataView.clearMarking()
+    };
 
-    donutChart(size, data, mod);
-
-    context.signalRenderComplete();
+    return donutState;
 }
 
 /**
